requirements.txt
flask
flask-sqlalchemy
python-dotenv
wtforms
flask-login
pydantic
pydantic-ai
httpx
logfire
PyJWT
